requirements.txt
keras==2.12.0
matplotlib==3.7.1
numpy==1.23.5
streamlit==1.22.0
tensorflow==2.12.0
pillow